Fixed imports for self energies having moved to processing.
