+++ modelforge/dataset/ani2x.py
@@ -164,7 +164,7 @@ class ANI2xDataset(HDF5Dataset):
 
     @property
     def atomic_self_energies(self):
-        from modelforge.potential.utils import AtomicSelfEnergies
+        from modelforge.potential.processing import AtomicSelfEnergies
 
         return AtomicSelfEnergies(energies=self._ase)
