+++ modelforge/dataset/spice114.py
@@ -200,7 +200,7 @@ class SPICE114Dataset(HDF5Dataset):
 
     @property
     def atomic_self_energies(self):
-        from modelforge.potential.utils import AtomicSelfEnergies
+        from modelforge.potential.processing import AtomicSelfEnergies
 
         return AtomicSelfEnergies(energies=self._ase)
