+++ modelforge/dataset/spice114openff.py
@@ -210,7 +210,7 @@ class SPICE114OpenFFDataset(HDF5Dataset):
 
     @property
     def atomic_self_energies(self):
-        from modelforge.potential.utils import AtomicSelfEnergies
+        from modelforge.potential.processing import AtomicSelfEnergies
 
         return AtomicSelfEnergies(energies=self._ase)
